feat: added the movie info card
