+++ src/Styles/MovieInfoCard.css
@@ -1,0 +1,81 @@
+.movie-info-card {
+    display: flex;
+    flex-direction: column;
+    background: #1a1a1a;
+    border-radius: 16px;
+    overflow: hidden;
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    box-shadow: 0 15px 35px rgba(0, 0, 0, 0.5);
+    max-width: 800px;
+    margin: 20px auto;
+    color: #ffffff;
+}
+
+.movie-info-poster {
+    width: 100%;
+    flex-shrink: 0;
+}
+
+.movie-info-poster img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    display: block;
+}
+
+.movie-info {
+    padding: 24px;
+    background: linear-gradient(to bottom, #292929, #000000);
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+}
+
+.movie-info h1 {
+    margin: 0;
+    font-size: 1.8rem;
+    color: #ffffff;
+}
+
+.movie-info-date {
+    color: #aaa;
+    font-size: 0.9rem;
+    margin-bottom: 8px;
+}
+
+.movie-info-overview {
+    line-height: 1.6;
+    color: #e0e0e0;
+    font-size: 1rem;
+}
+
+.movie-info-stats {
+    display: flex;
+    gap: 15px;
+    margin-top: auto;
+    padding-top: 15px;
+    border-top: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.movie-info-stats p {
+    background: rgba(255, 255, 255, 0.05);
+    padding: 4px 12px;
+    border-radius: 8px;
+    font-size: 0.85rem;
+    color: #f1f1f1;
+}
+
+@media (min-width: 768px) {
+    .movie-info-card {
+        flex-direction: row; 
+        align-items: stretch;
+    }
+
+    .movie-info-poster {
+        width: 300px; 
+    }
+
+    .movie-info h1 {
+        font-size: 2.2rem;
+    }
+}
